feat: add tab navigator and redux navigation
