--- app/homeTab/views/HomeTabScreenOne.js
@@ -1,27 +1,0 @@
-'use strict'
-import React from 'react'
-import { View, Text, TouchableOpacity } from 'react-native'
-export default class HomeTabScreenOne extends React.Component {
-  render(){
-    return(
-      <View style={{
-        flex:1,
-        backgroundColor:'red',
-        alignItems:'center',
-        justifyContent:'center'
-      }}>
-        <Text>{ 'Home Screen One' }</Text>
-        <TouchableOpacity
-          onPress={ () => this.props.navigation.navigate('HomeTabScreenTwo') }
-          style={{
-            padding:20,
-            borderRadius:20,
-            backgroundColor:'yellow',
-            marginTop:20
-          }}>
-          <Text>{'Go to next screen this tab'}</Text>
-        </TouchableOpacity>
-      </View>
-    )
-  }
-}
