+++ app/tabs/groupTab/components/GroupTabNavigation.js
