+++ app/tabs/groupTab/components/GroupTabScreenOne.js
